[harpoon] Initial setup to try out Harpoon (v2)
Added a plugin file for harpoon along with a _starter set_ of keybinds
to account for basic operations. In initial tests, it seems that "add"
and "remove" will use all files for a given buffer. This can cause
confusion for me as I usually keep multiple files in a buffer at once.
May follow up to investigate a better way later.